provision/docker: prevent wrong expansion on app image removal
Whenever removing an application, we use a regular expression for
removing its image from history, but the regular expression was not
accurate enough, which led to wrong images being deleted.

The test case in the code presents a case where the failures always
happened: if I'm willing to remove an application named "myapp", tsuru
will remove the app and use a regex to remove all of its image, and the
regex would be "tsuru/app-myapp" (assuming that no private registry is
configured, otherwise the regex will honor the registry), i.e.: tsuru
would remove images from any app whose name _starts_ with myapp. In
order to fix that, the regex is now "tsuru/app-myapp:v\d+".

Fix #1302.