provision/docker: update the container in the database before start
I could add a mock test for ensuring the pipeline order, but I don't
feel like this is a good test, maybe if we could somehow require
startContainer to query the database using the container ID (without
manually changing the Forward function in the test). If others think
that this is a must-exist test, I might take this approach.

Fix #1303.