api/server: bump to 0.12.4-galeb
This is supposed to be an internal release as part of our migration to
the new version of Galeb, but it's probably going public next week. This
is the first time (and hopefully the last) we run "unreleased code" in
production. We're not going to make a package for this release, but we
should get 0.12.4 with the proper fixes at some point in the next week,
and then we should get the 0.13.0 release with galeb, probably in the
week after.