api,app: initial experiment for more efficient log consumption
This is a work in progress and I'll add some benchmarks latter as I work
on this problem.

This initial change makes writes to mongodb in bulks of 100 log messages
or every 500 milliseconds, whatever comes first. Also one goroutine is
started for each application sending logs and log messages are sent from
the websocket handler to the consumer goroutine using a buffered channel
with 10000 entries. This should be enough to prevent the handler from
blocking as the most expensive operation is writing to mongodb and this
should only happen every 100 messages.