service/endpoint: improve error reporting in Create
At the moment, we're going to use an specific error for duplicate
instances only.

This is different from the other error declared in the
service_instance.go file. ErrInstanceAlreadyExistsInAPI is an internal
error in tsuru: if the isntance doesn't exist in tsuru but exists in the
API, there's nothing that the user can do.