service: ignore errors from UnbindApp when the API reports 404
We have an internal service where users are able to remove the service
instance directly, which means that tsuru may remain in an inconsistent
state. If the remote API returns 404, we're fine to just ignore the
failure in UnbindApp.

It's important to keep in mind that we're not ignoring it on creative
actions: BindApp still reports this error. So if users find that the
service instance they're trying to bind doesn't exist in the API, they
have only one option: running tsuru service-remove.