docs/releases: link the tsurud command for permission migration
We took this approach instead of using the script that we used
internally, because we don't want to force users to be able to compile a
Go program and run it in a machine that have access to the database.
There's already a machine with a compiled binary that have access to the
database: the API machine.

We needed to add support for optional migrations in order to accomplish
that, but now that we have optional migrations, it was an easier fix.