all: vendor all the things
We're not going to fight this new trend, but we're giving up on Godeps.
govendor does not include unused packages and tests when vendoring a
dependency. Since the Go team considers the fact that go test ./... run
tests for vendored dependencies a feature, we're going to avoid
vendoring tests.

Some changes were needed in Travis as well:

 - Go 1.4 doesn't support vendoring (we could do some magic, but I
   preferred to just drop it)
 - Go 1.5 requires the GO15VENDOREXPERIMENT environment variable
 - `make get` call is no longer needed (it doesn't even exist anymore)

Like the last item described, we don't have a target for downloading
dependencies anymore, because they're vendored :-) We also don't need
godep anymore \o/