provision/docker: ensure routes are up-to-date after changing units
It was possible for failures during communication with the router to
leave incorrect container addresses in it.

To ensure that the router correctly map entries in tsuru we'll always
call the rebuild routes operation to update the router based on what's
on tsuru database.

If the rebuild operation fails (e.g. router is still down after some
partially successful operation) a task will be added to monsterqueue.
This task will try to execute the rebuild operation repeatedly. It will
only end when a successful rebuild is executed or when the application
is removed from tsuru.