iaas/ec2: support all fields in ec2.RunInstancesInput
There's a lot of ugly magic going on, specially when we need to load a
slice of pointers (like in []*ec2.BlockDeviceMapping). We have the
option to explicitly handle each field, but then it's not extensible.
Anyway, feedback is very welcome!

I choose to handle monitoring explicitly because I think that it doesn't
make sense to force tsuru users to provide monitoring='{"Enabled":true}'
when they can simply provide monitoring-enabled=true.