provision/docker: improve error reporting on ambiguous containers
There's a new debug log saying which containers were found, and it now
includes the original ID/name used in the query. I've seen this error
happening more often than not, and I wanted to have more information for
debugging (just like we did in the past with
provision.UnitNotFoundError).