all: events now save allowed permissions during creation
This makes it a lot easier to filter permissions for each event when
listing and showing them.
Also we have more infromation about who will be able to see the event
when it's created. This is specially useful for delete events as the
target will no longer exist when showing the event.