fix race condition when registering a build container
It was possible for tsuru to receive a Register call targeting a build
container that still had a "created" status as the "building" status was
only set after the container started. This would cause tsuru to fail
later when trying to update the status of a building container (which is
forbidden by the query in Container#SetStatus.