all: directly communication between app and router without provisioner
To allow this change a new package was added to deal with the routes
rebuild logic. The app is now responsible for calling/enqueueing a
routes rebuild operation after calls to the provisioner that may change
routes.
Also calls to methods router methods like Addr(), SetCname() and
UnsetCname() are made directly from the app package.