api,provision: generic nodecontainers implementation to any provisioner
Now provisioners willing to support node containers need only to
implement a single interface method.

This is a work in progress, tests still are running in the
provision/docker package, but testing the implementation in the api
package, my next commit will split the tests and move them to the
correct place.