provision/docker: uses docker:tls:root-path to set nodes cert
When docker:tls:root-path is configured and the node does not
have an tls configuration already set, use the files provided
on the configuration file to set up the node tls.

This was done to keep compatibility with the tsuru installer and
should be removed once #1471 is done and a new version of the installer
is released.