iaas: choose default iaas from configured iaases
This commit changes the way tsuru chooses a default iaas when none
is specified.
First it will look for the `iaas:default` configuration entry.
If not found and only one iaas is configured, this one will be used.
If multiple iaases are configured and one of them is ec2,
ec2 will be used.
Else, it fails and returns an error.

Closes #1259