all: more reliable SetUnitStatus implementation for multiple provisioners
The previous implementation was too error prone because we tried to
update the unit status on all provisioners. This was NotFound responses
couldn't be trusted and BS would be unable to correctly delete
containers that are not supposed to exist.

Now we relly on the logic previously contained in the node healer
package to identify the node based on its containers or its addresses.
This is not ideal as it's not fast to search for the container on each
provisioner but this can be improved in the future. By updating BS to
send labels associated to the containers we can implement a fast and
reliable way to identify the which node is being reported.