provision/docker: scheduler choose container for removal from any host
Choosing from any host is a fallback behavior for when all existing
containers are on nodes that aren't registered in tsuru anymore. This
situation can happen if the nodes were removed and its units were not
moved/rebalanced to other nodes.

More tests were added to validate this new behavior.