provision,app: use pool constraints for default router
With this commit, the first router from the pool constraint
values will be used as the default value for that pool. We will discard
this if the first value has the character "*" or the constraint is a
blacklist. The global default value is used as fallback in those cases
and when there is no constraint for the pool.