service: use real app implementation on bind tests
Using FakeApp here was a really bad idea. It's not expected to ever
exist a second type of app and since the behavior of service and apps
depend so much on one another we were really only testing part of the
behavior.

FakeApp should only be used if there's no other option to avoid such
situations.