all: rework how service variables are associated to app
Instead of mixing them with user set variables we know store all service
envs as separated entries in the database. Also, they are stored as a
list, whereas user set variables are a map. This was done to preserve
the previous behavior of not losing data when services inject
conflicting variables.

The special env TSURU_SERVICES is now dinamically generated based on the
list of service envs associated to the app and no internal logic depends
on its value anymore. Another commit will provide the code necessary to
migrate values from the TSURU_SERVICES env to the new structure for
existing tsuru instalations.