api: use gorilla websocket for app-shell with ping/pong and deadlines
According to golang.org/x/net/websocket docs the gorilla websocket
package is recomended as being more maintained and containing more
features.

Using it we're able to add read deadlines with ping/pong messages being
used to extend this deadline. This allows tsuru to detect when a client
lost connection with the api server and finish the event and the remote
connection to the container.