healer,provision: rotate slices with resources for healing
This prevents a single set of nodes or containers from starving other
resources waiting to be healed. Today this can happen due to events rate
limiting only allowing a fixed number of healing events to run
concurrently. If the chosen resources fail to heal the healer process
would select them again as the first they would still be the first
elements in the collection.