event: ensure updates to lockUpdater internal set are atomic and synced
Before this commit the added test case would break due to the 10 buffer
in both addCh and removeCh channels. These buffers allowed the
possibility of a removeCh message to arrive before a addCh message for
the same ID, causing the message to never be removed from the internal
ID Set.

Simply removing the buffer could potentially cause performance issues
as every operation on events (New, Done, etc) would require the
lockUpdater goroutine to be running and unblocked. Long mongodb update
operations could delay the processing of seemingly unreleated events for
a long period of time.

This new approach uses a mutex based approach to update an internal
event ID Set and operations to the Set are made on the same goroutine as
the caller. The lockUpdater goroutine simply reads a copy of the
internal Set before updating the entries in the database ensuring
updates do not impact in Set operations.