app: handle race during default plan creation
The planService DefaultPlan() method is racy. When two apps are created
in parallel there is a change that we try to insert the "autogenerated"
plan multiple times, failing with: `Error: plan already exists`.

This commit adds a check to validate that after a "plan already exists"
error we are able to retrieve a default plan and return it.

This was caught by our integration tests which create apps in parallel.