provision/kubernetes: reduce number of kubernetes API reqs in Units() call
This change results in an improved performance for the app-list command
and was motivated by the analysis of a production profile with many
app-list requests happening.

Previously tsuru would make 1 query for Pods on the kubernetes API for
each App. Now a single query is made using the In operator in the
labelSelector.

Similarly each Node of each Pod would be queried with a Get() call, now
we also use the In operator on a fieldSelector.

We also reduced local MongoDB calls by calling it once per Pool instead
of once per App.