api,app,provision: single ExecuteCommand implementation in provisioners
Unifying all previous methods allow us to simplify the logic behind each
implementation. Now the differece between app-shell and app-run is
simply the presence of a Stdin reader in the ExecuteCommand call. This
change also removes some logic from inside each provisioner and moves it
to the `app` package, allowing it to be shared (selecting which units to
run and which commands to be executed).

Another benefit is that we now also have `app-shell --isolated` almost
for free.

Fixes #1827