all: increase all dial timeouts to 15 seconds
The motivation for this change was that the dial timeout includes the
time taken to resolve DNS names. And 5 seconds was too short to allow
recovery in case of packet losses.

By default, both Go and libc resolvers will issue both A and AAAA
queries. If one of these UDP packets is lost the resolver will wait for
its timeout value before returning (even if one of the queries already
finished successfully). The default timeout in the resolver is 5 seconds
with 2 attempts (per man resolv.conf). However, because our dial timeout
is also 5 seconds we would fail immediately without allowing the
resolver to run its course possibly returning a valid IP after 5 seconds
if only the AAAA packet was lost.

This change incresase dial timeout in all of our clients to 15 seconds,
this allows for 2 attempts of 5 seconds plus some spare time.