api: run shutdown handlers alongside server stop
This is needed because some shutdown handlers are required for allowing
a clean shutdown of the server. One suck example is the app-log -f
shutdown handler which will close existing log connections.

This commit also fix a race where if a stop signal is triggered before
the server is listening the shutdown handlers would run but the server
would still start and listen for connections.

Fixes #2050.