cmd: implements verbosity inside a custom RoundTripper
Prior to this change, the client object was responsible for providing
the request/response dumping according to the given verbosity level.
Commands using the new go-tsuruclient (generated from the spec) didn't
support the verbosity parameter because they do not use Client.

This change implements a custom RoundTripper that provides
request/response dumping according to a verbosity level. Commands
implemented with the new client simply have to pass down the http.Client
to the go-tsuruclient instance.