router/galeb: fix bug where an existing router would be deleted
During AddBackend calls started by the routes rebuild process, if an API
call failed with an error different from 409 there was the possibility
of tsuru completely removing the backend and all it's routes.

This commit fix this behavior and only removes existing resources if
we're dealing with a new backend.