app/image: always return a existing process on GetImageWebProcessName
Prior to this change, GetImageWebProcessName would return "web" if the
application had more than a single process, even if none of them was
actually called "web". This causes a issue on the Kubernetes provisioner
because it relies on the returned name to find some cluster resources,
cause the RoutableAddresses() method to fail.

This change makes sure that GetImageWebProcessName always return a
existing process name. If the application has more than a single process
it will either return "web", if one of them is called "web", or the
first process in alphabetical order.