app: enable filtering apps with specific platform versions
Prior to this change, users could not filter apps that were using a
specific plataform version (locked).

This change makes it possible to filter apps that are locked to a
specific platform version or are not locked to a specific version. Usage
is as follows:

`app list -p go` -> returns all apps using go (locked or not)
`app list -p go:v1` -> returns all apps that are using go locked to the
v1 version
`app list -p go:latest` -> return all apps that are not locked to a
specific go platform version.