api/server,docs/ref/config: support TLS certificates hot-reload
Prior to this change, it was needed to restart the web server to provide the
newer TLS certificates on disk. That's bad practice because involves downtime
of service.

This change allows that TLS certificates will be reloaded on runtime by the web
server. That reload process occurs ever than the certificates on disk are
different of the currently in-memory. The frequency of this process is defined
by "tls:auto-reload:interval" setting on the configuration file.