api/server: prevent to offer invalid TLS certificates (#2228)
* api/server: prevent to offer invalid TLS certificates

This change adds a new flag option on Tsuru API configuration file to
prevent it offers invalids certificates to its clients. In worst case,
the web server is gracefully shutdown to avoid other problems.

I can think a couple of problems that can occur when the certificates
served by Tsuru API to expire. For instance, one of these is related
between Big Sibling and Tsuru API. If BS can't talk with Tsuru API
(because certificate became invalid), the last one will think the node
(related to that BS) died and will heal it. Now imagine this for
many nodes... That's may be a big problem. Because of that, I think
it's safer validate certificates on Tsuru API before to offer them.