app: drop logs in bulk slice in case of mongodb failure
Emptying the bulk slice will cause the goroutine to keep consuming the
queue even in case of failures and will also prevent mongodb from
receiving a large bulk once it comes back online.
We also add a message to the log stream containing a warning of dropped
log message due to mongodb failure.