all: Ensure image versions are never reused and always incremented
It was possible for a failed build to reuse the same version multiple
times for each failed attempt. This happened because the version counter
was only incremeneted *after* the build step and just before the deploy
step.

The changes introduced here ensure we always increment the version
counter in the database before any other action is taken, this
incremented version is then propagated to the provisioner.